Long to int , deleting results

+++ EbayApi/src/main/java/com/project/pik/EbayApi/repositories/FoundResultRepository.java
@@ -16,5 +16,5 @@ public interface FoundResultRepository extends JpaRepository<FoundResult,Long>{
 	public List<FoundResult> findByOrder(Order order);
 	
 	@Transactional
-    Long deleteByFoundResultId(Long firstName);
+    Long deleteByFoundResultId(Integer firstName);
 }
